Use records instead of modules
Modules are actually unnecessary for what we want to do, although there is a
syntactic tradeoff when doing things via records - we no longer need to specify
"with type" constraints everywhere, and our "functor" definitions are simpler;
however, the types inside the records are more verbose, and what we end up with
is not very idiomatic OCaml, and lacks some tooling support.

Generally, we can do this for any module that does *not* define an inner type
which is itself parameterised - e.g. we can't do this for our Prob monad - and
does not rely on module inheritance to type-check correctly.

Slightly related is http://www.haskellforall.com/2012/05/scrap-your-type-classes.html