Implement the main Payout Table; this is now usable as a program
The results for the default case (no command-line arguments) are identical
to the original Python program and the run-time is similar compared to PyPy
which is itself many times faster than CPython. This is a very good sign since
this OCaml version has way more run-time error-checking than the original.